requirements.txt
numpy
Pillow
pikepdf
opencv-python
pyinstaller
numpy
fire
# python-docx
# docxtpl
